Move copy pasted Airnode stuff to separate folder
